requirements.txt
# requirements.txt
python-binance>=1.0.19
pandas>=1.3.0
cachetools>=5.0.0
ratelimit>=2.2.1
pydantic>=1.9.0
requests>=2.26.0
pytest>=7.0.0
flake8>=4.0.0
pytest-cov
pytest-asyncio==0.23.5  # Add for async tests
httpbin==0.7.0          # Add for connection pool testing
pydantic-settings>=2.0.3
python-binance>=1.0.19
pytest-asyncio>=0.23.0